moving status view files.
